refactor: move ReplacePackage to model-utils
