+++ model/utils/src/test/java/io/sundr/model/visitors/ReplacePackageTest.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package io.sundr.codegen;
+package io.sundr.model.visitors;
 
 import org.junit.Test;
